refactor(router): 라우터 및 페이지 임포트 경로 정리
- 라우터 파일에서 페이지 컴포넌트 임포트 경로를 간소화하여 가독성 향상
- 불필요한 주석 및 코드 정리
