+++ src/entities/order/data/index.ts
@@ -1,4 +1,2 @@
-// * Order Data Public API
-
 export { orderCardMock } from './orderCardMock'
 export { ORDER_FORM_PLACEHOLDER } from './orderContent'
