+++ src/entities/order/schema/index.ts
@@ -1,9 +1,8 @@
-// * Order Schema Public API
-
 export {
   orderFormSchema,
   receiverSchema,
   receiversSchema,
   type OrderFormData,
   type ReceiverData,
+  type CardData,
 } from './orderFormSchema'
