+++ src/entities/order/schema/orderFormSchema.ts
@@ -1,17 +1,30 @@
 import { z } from 'zod'
 import { VALIDATE_RULES } from '@/shared/lib/validateRules'
 
+// * 카드 데이터 타입
+export type CardData = {
+  id: number
+  thumbUrl: string
+  imageUrl: string
+  defaultTextMessage: string
+}
+
 // * 받는 사람 개별 스키마
 export const receiverSchema = z.object({
+  // * 받는 사람 이름
   name: z.string().min(1, VALIDATE_RULES.name.required?.errorMsg),
+
+  // * 받는 사람 전화번호
   phone: z
     .string()
     .min(1, VALIDATE_RULES.phone.required?.errorMsg)
     .regex(VALIDATE_RULES.phone.regex?.value, VALIDATE_RULES.phone.regex?.errorMsg),
+
+  // * 수량
   count: z.number().min(1, VALIDATE_RULES.quantity.custom?.errorMsg),
 })
 
-// * 받는 사람들 배열 스키마
+// * 받는 사람들 배열 스키마 (중복 전화번호 체크 포함)
 export const receiversSchema = z
   .array(receiverSchema)
   .max(10, '최대 10명까지만 추가할 수 있습니다.')
@@ -44,11 +57,18 @@ export const receiversSchema = z
     }
   })
 
-// * 주문하기 폼 스키마
+// * 주문하기 폼 스키마 정의
 export const orderFormSchema = z.object({
+  // * 카드 메시지
   cardMessage: z.string().min(1, VALIDATE_RULES.message.required?.errorMsg),
+
+  // * 보내는 사람
   sender: z.string().min(1, VALIDATE_RULES.name.required?.errorMsg),
+
+  // * 받는 사람들
   receivers: receiversSchema,
+
+  // * 선택된 카드
   selectedCard: z.object({
     id: z.number(),
     thumbUrl: z.string(),
