+++ src/entities/product/data/index.ts
@@ -1,4 +1,2 @@
-// * Product Data Public API
-
 export { productListMock } from './productListMock'
 export { categoryMock } from './categoryMock'
