+++ src/entities/product/index.ts
@@ -1,7 +1,2 @@
-// * Product Entity Public API
-
-// Data re-exports
 export * from './data'
-
-// Schema re-exports
 export * from './schema'
