+++ src/entities/product/schema/index.ts
@@ -1,4 +1,1 @@
-// * Product Schema Public API
-
 // 현재 상품 관련 스키마 없음
-// 필요시 여기에 추가
