+++ src/entities/user/data/index.ts
@@ -1,3 +1,1 @@
-// * User Data Public API
-
 export { LOGIN_CONTENT } from './loginContent'
