+++ src/entities/user/index.ts
@@ -1,7 +1,2 @@
-// * User Entity Public API
-
-// Data re-exports
 export * from './data'
-
-// Schema re-exports
 export * from './schema'
