+++ src/entities/user/schema/index.ts
@@ -1,3 +1,1 @@
-// * User Schema Public API
-
 export { loginSchema, type LoginFormData } from './loginSchema'
