+++ src/features/order/types.ts
@@ -1,15 +1,7 @@
 import type { Product } from '@/features/product/trend/types'
-import type { ReceiverData } from '@/entities/order'
+import type { ReceiverData, CardData } from '@/entities/order'
 
-export type CardData = {
-  id: number
-  thumbUrl: string
-  imageUrl: string
-  defaultTextMessage: string
-}
-
-// * ReceiverData는 entities/order에서 import
-export type { ReceiverData }
+export type { ReceiverData, CardData }
 
 export type OrderForm = {
   card_data: CardData
