+++ src/features/product/index.ts
@@ -1,12 +1,7 @@
-// * Product Feature Public API
-
-// Trend 관련 컴포넌트들
 export { Trend } from './trend/Trend'
 export { ProductItem } from './trend/ProductItem'
 export { TrendFilter } from './trend/TrendFilter'
 
-// Types
 export type { Product, TargetType, RankType } from './trend/types'
 
-// Filter 데이터
 export { FilterItems, RankItems } from './trend/filterList'
