+++ src/pages/Order.tsx
@@ -1,16 +1,13 @@
 import { Button, PageContainer, Typography } from '@/shared/ui'
-import type { Product } from '@/features/product/trend/types'
 import { productListMock } from '@/entities/product'
+import { type Product } from '@/features'
 import { useEffect, useState } from 'react'
 import { useNavigate, useParams } from 'react-router-dom'
 import { NotFound } from './NotFound'
 import styled from '@emotion/styled'
 import { theme } from '@/shared/styles/theme'
-import { OrderCardSection } from '@/features/order/OrderCardSection'
-import type { CardData } from '@/features/order/types'
+import { OrderCardSection, SenderSection, ReceiverSection, type CardData } from '@/features'
 import { orderCardMock } from '@/entities/order'
-import { SenderSection } from '@/features/order/SenderSection'
-import { ReceiverSection } from '@/features/order/ReceiverSection'
 import { ROUTE_PATH } from '@/app/Router'
 import { OrderFormProvider, useOrderForm } from '@/shared/contexts/order'
