+++ src/pages/index.ts
@@ -1,0 +1,5 @@
+export { Home } from './Home'
+export { Login } from './Login'
+export { MyPage } from './MyPage'
+export { NotFound } from './NotFound'
+export { Order } from './Order'
