+++ src/shared/config/index.ts
@@ -1,4 +1,1 @@
-// * Shared Config Public API
-
-// Banner 관련 설정
 export { BANNER_CONTENT } from './bannerContent'
