+++ src/shared/contexts/auth/index.ts
@@ -1,3 +1,1 @@
-// * Auth Context Public API
-
 export { AuthProvider, useAuth, type UserInfo } from './AuthContext'
