+++ src/shared/contexts/order/OrderFormContext.tsx
@@ -1,9 +1,8 @@
 import { createContext, useContext, useState, type ReactNode } from 'react'
 import { useForm, type UseFormReturn } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { orderFormSchema, type OrderFormData } from '@/entities/order'
+import { orderFormSchema, type OrderFormData, type CardData } from '@/entities/order'
 import { orderCardMock } from '@/entities/order'
-import type { CardData } from '@/features/order/types'
 
 // * 주문 폼 컨텍스트 타입
 type OrderFormContextType = {
