+++ src/shared/contexts/order/index.ts
@@ -1,3 +1,1 @@
-// * Order Form Context Public API
-
 export { OrderFormProvider, useOrderForm } from './OrderFormContext'
