+++ src/shared/lib/index.ts
@@ -1,10 +1,6 @@
-// * Shared Lib Public API
-
-// 유틸리티 함수들
 export { deleteCookie, getCookie, setCookie } from './cookieUtils'
 export { decodeUserInfo, encodeUserInfo } from './encodingUtils'
 export { validateValue } from './validateValue'
 export { VALIDATE_RULES } from './validateRules'
 
-// 타입들
 export type { Rule } from './validateValue'
