+++ src/shared/ui/index.ts
@@ -1,4 +1,3 @@
-// * 공통 컴포넌트 export
 export { PageContainer } from './PageContainer'
 export { Button } from './Button'
 export { Modal } from './Modal'
